requirements_backup.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via starlette
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
ci-info==0.3.0
    # via etelemetry
click==8.1.7
    # via
    #   nipype
    #   uvicorn
configobj==5.0.9
    # via fitz
configparser==7.1.0
    # via fitz
etelemetry==0.3.1
    # via nipype
fastapi==0.115.0
    # via -r requirements.in
filelock==3.16.1
    # via nipype
google-api-core==2.20.0
    # via google-api-python-client
google-api-python-client==2.147.0
    # via -r requirements.in
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via google-api-core
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   fitz
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via
    #   anyio
    #   requests
isodate==0.6.1
    # via rdflib
looseversion==1.3.0
    # via nipype
lxml==5.3.0
    # via
    #   prov
    #   pyxnat
networkx
    # via
    #   nipype
    #   prov
nibabel==5.2.1
    # via
    #   fitz
    #   nipype
nipype==1.8.6
    # via fitz
numpy
    # via
    #   fitz
    #   nibabel
    #   nipype
    #   pandas
    #   scipy
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   etelemetry
    #   nibabel
    #   nipype
pandas==2.2.3
    # via
    #   -r requirements.in
    #   fitz
pathlib==1.0.1
    # via pyxnat
pdf2image==1.17.0
    # via -r requirements.in
pillow==10.4.0
    # via
    #   -r requirements.in
    #   pdf2image
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
prov==2.0.1
    # via nipype
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.23.4
    # via pydantic
pydot==3.0.2
    # via nipype
pymupdf==1.24.10
    # via -r requirements.in
pymupdfb==1.24.10
    # via pymupdf
pyparsing==3.1.4
    # via
    #   httplib2
    #   pydot
    #   rdflib
python-dateutil==2.9.0.post0
    # via
    #   nipype
    #   pandas
    #   prov
pytz==2024.2
    # via pandas
pyxnat==1.6.2
    # via fitz
rdflib==6.3.2
    # via
    #   nipype
    #   prov
requests==2.32.3
    # via
    #   -r requirements.in
    #   etelemetry
    #   google-api-core
    #   pyxnat
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scipy
    # via
    #   fitz
    #   nipype
simplejson==3.19.3
    # via nipype
six==1.16.0
    # via
    #   isodate
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.38.6
    # via fastapi
traits==6.3.2
    # via nipype
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
uvicorn==0.31.0
    # via -r requirements.in
